requirements.txt
torch
transformers
fastapi
uvicorn
pydantic
scikit-learn
shap
sentencepiece
dataset
accelerate
